chore: extract build hooks handlers
