+++ libs/zephyr-agent/src/index.ts
@@ -28,6 +28,7 @@ export {
   is_zephyr_dependency_pair,
   type ZeDependencyPair,
   type ZephyrDependencies,
+  type ZephyrEngineOptions,
   readPackageJson,
 } from './zephyr-engine';
 export type { ZeResolvedDependency } from './zephyr-engine/resolve_remote_dependency';
